+++ module/Motion/Rotation.cpp
@@ -34,8 +34,8 @@ void Rotation::run()
   // 継続条件を満たしている間ループ
   while(isMetPostcondition(initLeftMileage, initRightMileage, leftSign, rightSign)) {
     // PWM値を設定する
-    int leftPwm = speedCalculator.calcLeftPwmFromSpeed();
-    int rightPwm = speedCalculator.calcRightPwmFromSpeed();
+    double leftPwm = speedCalculator.calcLeftPwmFromSpeed();
+    double rightPwm = speedCalculator.calcRightPwmFromSpeed();
 
     // モータにPWM値をセット
     Controller::setLeftMotorPwm(leftPwm);
